// frontend/src/client/models/Meal.ts
// tslint:disable
/**
 * Macro Macro API
 * An API that provides access to recipe and ingredient information, both user-generated and sourced from the USDA
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    AmountOfNamedMacros,
    AmountOfNamedMacrosFromJSON,
    AmountOfNamedMacrosFromJSONTyped,
    AmountOfNamedMacrosToJSON,
} from './';

/**
 * 
 * @export
 * @interface Meal
 */
export interface Meal {
    /**
     * 
     * @type {string}
     * @memberof Meal
     */
    uid: string;
    /**
     * 
     * @type {Array<AmountOfNamedMacros>}
     * @memberof Meal
     */
    foods: Array<AmountOfNamedMacros>;
}

export function MealFromJSON(json: any): Meal {
    return MealFromJSONTyped(json, false);
}

export function MealFromJSONTyped(json: any, ignoreDiscriminator: boolean): Meal {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'uid': json['uid'],
        'foods': (json['foods'] as Array<any>).map(AmountOfNamedMacrosFromJSON),
    };
}

export function MealToJSON(value?: Meal): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'uid': value.uid,
        'foods': (value.foods as Array<any>).map(AmountOfNamedMacrosToJSON),
    };
}